requirements.txt
redis~=5.2.0
aiogram~=3.15.0
python-dotenv~=1.0.1
emoji~=2.14.0
aiohttp~=3.10